stricter type

+++ tests/structural_analysis.space_truss.node.cpp
@@ -1,7 +1,5 @@
 ﻿#include <catch.hpp>
-#include "approx_equal.h"
 
-#include <sec21/units.h>
 #include <sec21/structural_analysis/node.h>
 #include <sec21/structural_analysis/member.h>
 #include <sec21/structural_analysis/space_truss.h>
@@ -10,8 +8,8 @@ TEST_CASE("add node to system", "[sec21][structural_analysis][space_truss][node]
 {
    using namespace sec21::structural_analysis;
 
-   using member_t = member<int, double>;
-   using node_t = node<2, int, double>;
+   using member_t = member<int, float>;
+   using node_t = node<2, int, float>;
    using space_truss_t = space_truss<node_t, member_t>;
 
    SECTION("add one node")
@@ -32,8 +30,8 @@ TEST_CASE("add node to system", "[sec21][structural_analysis][space_truss][node]
    SECTION("add two nodes with the same id")
    {
       auto sys = space_truss_t{};
-      auto n1 = add_node(sys, { 1 });
-      auto n2 = add_node(sys, { 1 });
+      auto n1 = add_node(sys, { 1u });
+      auto n2 = add_node(sys, { 1u });
 
       REQUIRE(static_cast<bool>(n1) == true);
       REQUIRE(n1.value() == 1);
